requirements.txt
slack_bolt
python-dotenv
pika==1.1.0